make parser a module
