+++ docs/sphinx_doc/zh_CN/source/index.rst
@@ -33,6 +33,7 @@ AgentScope 文档
    agentscope.models
    agentscope.agents
    agentscope.memory
+   agentscope.parsers
    agentscope.pipelines
    agentscope.service
    agentscope.rpc
